workspace/SRM 650/TaroJiroDividing.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Numerics;
using Debug = System.Diagnostics.Debug;
using StringBuilder = System.Text.StringBuilder;
public class TaroJiroDividing
{
    public int getNumber(int A, int B)
    {
        var S = new HashSet<int>();
        var T = new HashSet<int>();
        while (A % 2 == 0)
        {
            S.Add(A);
            A /= 2;
        }
        S.Add(A);
        while (B % 2 == 0)
        {
            T.Add(B);
            B /= 2;
        }
        T.Add(B);
        return S.Intersect(T).Count();
    }

    static public T[] Enumerate<T>(int n, Func<int, T> f) { var a = new T[n]; for (int i = 0; i < n; ++i) a[i] = f(i); return a; }
    static public void Swap<T>(ref T a, ref T b) { var tmp = a; a = b; b = tmp; }
}
static public class EnumerableEX
{
    static public string AsString(this IEnumerable<char> e) { return new string(e.ToArray()); }
    static public string AsJoinedString<T>(this IEnumerable<T> e, string s = " ") { return string.Join(s, e); }
}
// CUT begin
public class Tester: AbstractTester
{
    static public T[] Enumerate<T>(int n, Func<int, T> f) { var a = new T[n]; for (int i = 0; i < n; ++i) a[i] = f(i); return a; }
    public Random rand = new Random(0);
    public Tester()
    {
    }
}
// CUT end
